fix(release): release automation fixes and refactoring
- Fix promote_rc dry run message to use correct remote.
- Support spaces in metadata parsing and format commit SHA with space for GitHub autolinking.
- Trigger release publish workflow from create_rc workflow using workflow_call.
- Replace version markers during backport processing.
- Split release_test.py into separate files for each command.
- Refactor release tool BUILD targets to use py_library and glob.
