+++ tools/private/release/BUILD.bazel
@@ -3,32 +3,18 @@ load("@rules_python//python:defs.bzl", "py_binary", "py_library")
 package(default_visibility = ["//visibility:public"])
 
 py_library(
-    name = "changelog_news",
-    srcs = ["changelog_news.py"],
+    name = "release_lib",
+    srcs = glob(["*.py"]),
+    deps = [
+        "@dev_pip//packaging",
+    ],
 )
 
 py_binary(
     name = "release",
-    srcs = [
-        "__init__.py",
-        "complete_prepare.py",
-        "create_rc.py",
-        "create_release_branch.py",
-        "create_release_issue.py",
-        "determine_next_version.py",
-        "gh.py",
-        "git.py",
-        "prepare.py",
-        "process_backports.py",
-        "promote_rc.py",
-        "release.py",
-        "release_issue.py",
-        "shell.py",
-        "utils.py",
-    ],
+    srcs = ["release.py"],
     main = "release.py",
     deps = [
-        ":changelog_news",
-        "@dev_pip//packaging",
+        ":release_lib",
     ],
 )
